[slimtensor] integration into backend
Differential Revision: [D90606409](https://our.internmc.facebook.com/intern/diff/D90606409/)

[ghstack-poisoned]

+++ backends/cuda/runtime/guard.h
@@ -19,8 +19,8 @@ namespace executorch::backends::cuda {
 using executorch::runtime::Error;
 using executorch::runtime::Result;
 
-// Type alias for device index
-using DeviceIndex = int32_t;
+// Signed device index type matching DeviceIndex in slim tensor library
+using DeviceIndex = int8_t;
 
 /**
  * Set the current CUDA stream for the specified device.
